Implement standalone dev-server using @web/dev-server as library
This will let us run two servers from the same command, plus we get
type checking for our plugins.

This also hard codes the dev server port to 5415 (an arbitrary unique
number).
