+++ packages/lit-dev-content/package.json
@@ -20,14 +20,13 @@
     "fonts:manrope": "rm -rf site/fonts temp && mkdir -p site/fonts && git clone https://github.com/sharanda/manrope.git temp/manrope && cd temp/manrope && git checkout 9ffbc349f4659065b62f780fe6e9d5a93518bd95 && cp fonts/web/*.woff2 ../../site/fonts/ && cd ../.. && rm -rf temp",
     "dev:build:site": "rm -rf _dev && ELEVENTY_ENV=dev PLAYGROUND_SANDBOX=. eleventy",
     "dev:build:site:watch": "rm -rf _dev && ELEVENTY_ENV=dev PLAYGROUND_SANDBOX=. eleventy --watch",
-    "dev:serve": "web-dev-server --root-dir=_dev --node-resolve --watch --open --preserve-symlinks",
+    "dev:serve": "node ../lit-dev-tools-esm/lib/dev-server.js --open",
     "format": "../../node_modules/.bin/prettier \"**/*.{ts,js,json,html,css,md}\" --write"
   },
   "devDependencies": {
     "@11ty/eleventy": "^0.12.1",
     "@11ty/eleventy-navigation": "^0.3.2",
     "@rollup/plugin-node-resolve": "^13.0.4",
-    "@web/dev-server": "^0.1.6",
     "clean-css": "^5.1.0",
     "codemirror": "^5.58.2",
     "eleventy-plugin-nesting-toc": "^1.1.0",
