+++ packages/lit-dev-tools-esm/package.json
@@ -15,6 +15,7 @@
     "@lit/ts-transformers": "^1.0.0",
     "@types/ansi-escape-sequences": "^4.0.0",
     "@types/prettier": "^2.3.2",
+    "@web/dev-server": "^0.1.6",
     "ansi-escape-sequences": "^6.2.0",
     "chokidar": "^3.5.2",
     "lit-dev-server": "^0.0.0",
